rename com to io
