+++ artmaker/src/main/java/io/artmaker/actions/ArtMakerAction.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.artmaker.actions
+package io.artmaker.actions
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
