+++ artmaker/src/main/java/io/artmaker/composables/ArtMakerControlMenu.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.artmaker.composables
+package io.artmaker.composables
 
 import android.annotation.SuppressLint
 import android.graphics.BitmapFactory
@@ -60,11 +60,11 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.core.os.BuildCompat
-import com.artmaker.actions.ArtMakerAction
+import io.artmaker.actions.ArtMakerAction
 import io.fbada006.artmaker.R
-import com.artmaker.models.ArtMakerConfiguration
-import com.artmaker.state.ArtMakerUIState
-import com.artmaker.utils.ColorUtils
+import io.artmaker.models.ArtMakerConfiguration
+import io.artmaker.state.ArtMakerUIState
+import io.artmaker.utils.ColorUtils
 import com.google.modernstorage.photopicker.PhotoPicker
 
 private const val IMAGE_PICKER_MAX_ITEMS = 1
