+++ artmaker/src/main/java/io/artmaker/composables/StrokeWidthSlider.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.artmaker.composables
+package io.artmaker.composables
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
@@ -32,10 +32,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
-import com.artmaker.actions.ArtMakerAction
+import io.artmaker.actions.ArtMakerAction
 import io.fbada006.artmaker.R
-import com.artmaker.models.ArtMakerConfiguration
-import com.artmaker.state.ArtMakerUIState
+import io.artmaker.models.ArtMakerConfiguration
+import io.artmaker.state.ArtMakerUIState
 
 /**
  * This is a popup that displays the ArtMakerStrokeWidthSlider and accompanying title.
