+++ artmaker/src/main/java/io/artmaker/utils/ExportHelper.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.artmaker.utils
+package io.artmaker.utils
 
 import android.content.Context
 import android.content.Intent
