+++ demo/src/main/java/com/artmaker/demo/MainActivity.kt
@@ -20,7 +20,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
-import com.artmaker.ArtMaker
+import io.artmaker.ArtMaker
 import com.artmaker.demo.ui.theme.ArtMakerTheme
 
 class MainActivity : ComponentActivity() {
